API change in libcore

+++ include/core/Base64.h
@@ -52,7 +52,7 @@ namespace jm
           \return The output buffer. The caller of the method takes over the array and has to clean
           it up by himself.
           */
-         static uint8* encode(const uint8* data, Integer& length);
+         static uint8* encode(const uint8* data, int64& length);
 
          /*!
           \brief Decodes the data stream with the BASE64 algorithm.
