+++ include/core/Vertex2.h
@@ -195,7 +195,7 @@ namespace jm
     \brief Implementation of the operator *
     */
    DllExport
-   const Vertex2 operator*(const Integer& i, const Vertex2& v);
+   const Vertex2 operator*(const int64& i, const Vertex2& v);
 
    /*!
     \brief Implementation of the operator -=
